+++ src/bin/vfilter.rs
@@ -8,7 +8,7 @@
 use std::ops::{BitXor, BitXorAssign};
 
 use anyhow::Result;
-use clap::Parser;
+use clap::{ArgGroup, Parser};
 use common_traits::{DowncastableFrom, UpcastableFrom};
 use dsi_progress_logger::*;
 use epserde::ser::Serialize;
@@ -27,18 +27,25 @@ use value_traits::slices::SliceByValueMut;
 
 #[derive(Parser, Debug)]
 #[command(about = "Creates a VFilter and serializes it with ε-serde", long_about = None)]
+#[clap(group(
+            ArgGroup::new("input")
+                .required(true)
+                .multiple(true)
+                .args(&["filename", "n"]),
+))]
 struct Args {
     /// The number of keys. If no filename is provided, use the 64-bit keys
     /// [0..n).
-    n: usize,
+    #[arg(short, long)]
+    n: Option<usize>,
+    /// A file containing UTF-8 keys, one per line. At most N keys will be read.
+    #[arg(short, long)]
+    filename: Option<String>,
     /// An optional name for the ε-serde serialized function.
     filter: Option<String>,
     /// The number of bits of the hashes used by the filter.
     #[arg(short, long, default_value_t = 8)]
     bits: usize,
-    #[arg(short, long)]
-    /// A file containing UTF-8 keys, one per line. At most N keys will be read.
-    filename: Option<String>,
     /// Use this number of threads.
     #[arg(short, long)]
     threads: Option<usize>,
@@ -140,8 +147,11 @@ fn set_builder<W: Word + BinSafe, D: BitFieldSlice<W> + Send + Sync, S, E: Shard
     let mut builder = builder
         .offline(args.offline)
         .check_dups(args.check_dups)
-        .expected_num_keys(args.n)
         .eps(args.eps);
+    if let Some(n) = args.n {
+        builder = builder.expected_num_keys(n);
+    }
+
     if let Some(seed) = args.seed {
         builder = builder.seed(seed);
     }
@@ -187,9 +197,9 @@ where
     let mut pl = ProgressLogger::default();
     #[cfg(feature = "no_logging")]
     let mut pl = Option::<ConcurrentWrapper<ProgressLogger>>::None;
-    let n = args.n;
 
     if let Some(filename) = &args.filename {
+        let n = args.n.unwrap_or(usize::MAX);
         let builder = set_builder(VBuilder::<W, Box<[W]>, S, E>::default(), &args);
         let filter = if args.zstd {
             builder.try_build_filter(ZstdLineLender::from_path(filename)?.take(n), &mut pl)?
@@ -201,6 +211,7 @@ where
             unsafe { filter.store(filename) }?;
         }
     } else {
+        let n = args.n.unwrap();
         let builder = set_builder(VBuilder::<W, Box<[W]>, S, E>::default(), &args);
         let filter =
             builder.try_build_filter(FromCloneableIntoIterator::from(0_usize..n), &mut pl)?;
@@ -234,9 +245,9 @@ where
     let mut pl = ProgressLogger::default();
     #[cfg(feature = "no_logging")]
     let mut pl = Option::<ConcurrentWrapper<ProgressLogger>>::None;
-    let n = args.n;
 
     if let Some(filename) = &args.filename {
+        let n = args.n.unwrap_or(usize::MAX);
         let builder = set_builder(VBuilder::<W, BitFieldVec<W>, S, E>::default(), &args);
         let filter = if args.zstd {
             builder.try_build_filter(
@@ -252,6 +263,7 @@ where
             unsafe { filter.store(filename)? };
         }
     } else {
+        let n = args.n.unwrap();
         let builder = set_builder(VBuilder::<W, BitFieldVec<W>, S, E>::default(), &args);
         let filter = builder.try_build_filter(
             FromCloneableIntoIterator::from(0_usize..n),